hash: fix compiler warnings
In file included from tools/common/hash.c:1:0:
./tools/../common/hash.c:123:12: warning: ‘hw_dummy_init’
defined but not used [-Wunused-function]
 static int hw_dummy_init(struct hash_algo *algo, void **ctxp)
            ^
./tools/../common/hash.c:129:12: warning: ‘hw_dummy_update’
defined but not used [-Wunused-function]
 static int hw_dummy_update(struct hash_algo *algo, void *ctx,
            ^
./tools/../common/hash.c:136:12: warning: ‘hw_dummy_finish’
defined but not used [-Wunused-function]
 static int hw_dummy_finish(struct hash_algo *algo, void *ctx, void *dest_buf,

 In file included from drivers/crypto/img-hash.c:9:0:
drivers/crypto/img-hash.c: In function 'img_hash_xmit':
./arch/mips/include/asm/io.h:186:29: warning: cast to pointer from integer of
different size [-Wint-to-pointer-cast]
 #define __raw_readb(addr) (*(volatile unsigned char *)(addr))
                             ^
drivers/crypto/img-hash.c:181:25: note: in expansion of macro '__raw_readb'
   *(bufp + readcount) = __raw_readb(addr + readcount + offset);

+++ drivers/crypto/img-hash.c
@@ -178,7 +178,8 @@ static int img_hash_xmit(u64 len, u64 addr)
 	}
 	/* and then write the remaining bytes */
 	for (readcount = 0; readcount < remainder; readcount++) {
-		*(bufp + readcount) = __raw_readb(addr + readcount + offset);
+		*(bufp + readcount) = __raw_readb((u32)(addr +
+						readcount + offset));
 		debug("in buffer : %u\n", buf);
 		debug("readcount : %d, addr: %d, offset: %d\n",
 		      (int)readcount, (int)addr, (int)offset);
